+++ app/Providers/AppServiceProvider.php
@@ -26,6 +26,7 @@
 namespace Gamify\Providers;
 
 use Illuminate\Support\ServiceProvider;
+use Illuminate\Validation\Rules\Password;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -36,6 +37,16 @@ class AppServiceProvider extends ServiceProvider
 
     public function boot(): void
     {
-        //
+        Password::defaults(function () {
+            $rule = Password::min(8)
+                ->letters()
+                ->mixedCase()
+                ->numbers()
+                ->symbols();
+
+            return $this->app->isProduction()
+                ? $rule->uncompromised()
+                : $rule;
+        });
     }
 }